pyproject.toml
[project]
name = "rag-commander"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.57.0",
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "google-auth-oauthlib>=1.2.3",
    "jupyter>=1.1.1",
    "langchain==1.0.3",
    "langchain-anthropic>=1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.0.2",
    "langchain-experimental>=0.0.42",
    "langchain-google-genai>=3.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai==1.0.1",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.0",
    "markdown>=3.9",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pdfminer>=20191125",
    "pdfplumber>=0.11.7",
    "perplexityai>=0.17.1",
    "pgvector>=0.4.1",
    "pi-heif>=1.1.1",
    "pillow>=11.3.0",
    "pip>=25.3",
    "psutil>=7.1.2",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.5",
    "pypdf>=4.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.7.2",
    "selenium>=4.37.0",
    "sentence-transformers>=5.1.1",
    "tavily-python>=0.7.12",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "unstructured>=0.18.15",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]  # src 하위의 패키지 전부(예: tools, utils 등). 각 폴더에 __init__.py 필요

[tool.uv]
override-dependencies = ["bitsandbytes==0.0.0"]
package = true  
